MODIFIED: Added changes made with hotfix a+b MODIFIED: Fixed image picker bug MODIFIED: Fixed new raid sheet layout MODIFIED: Fixed patch level update notification

+++ lib/script/config.js.php
@@ -50,7 +50,7 @@
     $Settings->closeCursor();
 ?>
 
-var g_SiteVersion = <?php echo intval($_REQUEST["version"]) ?>;
+var g_SiteVersion = <?php echo floatval($_REQUEST["version"]) ?>;
 var g_BannerLink = "<?php echo $Site; ?>";
 var g_TimeFormat = <?php echo $TimeFormat; ?>;
